requirements.txt
certifi==2022.12.7
kafka-python==2.0.2
wincertstore==0.2
requests~=2.30.0